pyproject.toml
[project]
name = "data-nadhi-temporal-worker"
version = "0.1.0"
description = "A Python SDK for Data Nadhi"
requires-python = ">=3.11"
authors = [
    { name = "Nishaanth S P", email = "nishaanth.sp@dataarena.space" }
]
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "temporalio==1.18.1",
    "minio>=7.0.0"
]

[project.optional-dependencies]
main = [
    "pymongo>=4.0.0",
    "redis>=4.0.0"
]
transformation = []
destination = [
    "pymongo>=4.0.0",
    "redis>=4.0.0",
    "cryptography",
    "slack_sdk"
]
dev = [
    "pymongo>=4.0.0",
    "redis>=4.0.0",
    "cryptography",
    "slack_sdk",
    "pytest>=7.0.0",
    "ruff>=0.0.291"
]

[tool.ruff]
line-length = 88
target-version = "py311"
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "W", "N", "UP", "RET", "LOG007", "PLR2004"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []